Left bar change + JS change
Changed the left bar to open and close with a button, hamburger to open and x-icon to close.

Made the calendar open and close with an arrow.

Made the main task list take up the whole screen, but its now scrollable, the project list is always visible even if you open the calandar and have a full list.

Startet on JS-code: Added uniqe id for each list element so it can be styled in the pop-up window, going to add the ability to give single tasks a done status.

Co-Authored-By: vemundkristiansen <vemundkristiansen@users.noreply.github.com>